% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{build_bins}
\alias{build_bins}
\title{Build a bins GRanges object.}
\usage{
build_bins(bin_size = 10000, genome = "mm9")
}
\arguments{
\item{bin_size}{Bin size. 10000 by default.}

\item{genome}{Genome used. mm9 by default. Valid values: mm9, hg38}
}
\value{
A GRanges object
}
\description{
Build a GRanges of bins of a given size, for a specific genome. Supported
genomes (required for the package): mm9, hg38.
}
