% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bed}
\alias{bw_bed}
\title{Build a scored GRanges object from a BED file.}
\usage{
bw_bed(bwfile, bedfile, per.locus.stat = "mean", aggregate.by = NULL)
}
\arguments{
\item{bwfile}{BigWig file to be summarized.}

\item{bedfile}{BED file to intersect with the BigWig file.}

\item{per.locus.stat}{Aggregate per locus}

\item{aggregate.by}{Statistic to aggregate per group. If NULL, values are not aggregated. This is the behavior by default.}

\item{keep.name}{Keep the name of specific locus (adds one mdata col to the GRanges object). True by default.}
}
\description{
Build a scored GRanges object from a bigWig file and a BED file.
The aggregating function (per locus) can be min, max, sd, mean.
}
