% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{bw_bed_diff_analysis}
\alias{bw_bed_diff_analysis}
\title{Run DESeq2 analysis on bed file}
\usage{
bw_bed_diff_analysis(
  bwfiles_c1,
  bwfiles_c2,
  bedfile,
  label_c1,
  label_c2,
  estimate_size_factors = FALSE,
  as_granges = FALSE,
  length_factor = 1000,
  shrink = TRUE,
  p_cutoff = 0.05
)
}
\arguments{
\item{bwfiles_c1}{Path or array of paths to the bigWig files for first condition.}

\item{bwfiles_c2}{Path or array of paths to the bigWig files for second condition.}

\item{bedfile}{BED file for locus specific analysis.}

\item{label_c1}{Condition name for condition 1.}

\item{label_c2}{Condition name for condition 2.}

\item{estimate_size_factors}{If TRUE, normal DESeq2 procedure is done. Set it
to true to analyze non-MINUTE data.}

\item{as_granges}{If TRUE, returns a GRanges object. This is useful when
loci coordinates are relevant.}

\item{length_factor}{Value to multiply read counts by. Ideally, this factor
should approximate locus_length / fragment_length to simulate actual read
counts from bigWig coverage.}

\item{shrink}{If TRUE, apply apelgm method for low counts shrinkage}

\item{p_cutoff}{Pvalue cutoff to find a result significant}
}
\value{
a DESeqResults object as returned by DESeq2::results function
}
\description{
Runs a DESeq2 analysis on a set of loci specified in a BED file.
The particularity of this analysis is that it skips the estimateSizeFactors
step by default, because this is accounted for in the scaling step of
MINUTE-ChIP samples.
}
