% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bins}
\alias{bw_bins}
\title{Build a binned-scored GRanges object from a bigWig file}
\usage{
bw_bins(bwfile, stat = "mean", bsize = 10000, genome = "mm9")
}
\arguments{
\item{bwfile}{BigWig file to be summarized.}

\item{stat}{Aggregating function. Mean by default. Choices: min, max, sd, mean.}

\item{bsize}{Bin size. Default 10000.}

\item{genome}{Genome. Available choices are mm9, hg38.}
}
\description{
Build a binned-scored GRanges object from a bigWig file. The aggregating
function can be min, max, sd, mean.
}
