% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{bw_bins_diff_analysis}
\alias{bw_bins_diff_analysis}
\title{Run DESeq2 analysis on genome-wide bins}
\usage{
bw_bins_diff_analysis(
  bwfiles_c1,
  bwfiles_c2,
  label_c1,
  label_c2,
  bin_size = 10000,
  genome = "mm9"
)
}
\arguments{
\item{bwfiles_c1}{Path or array of paths to the bigWig files for first condition.}

\item{bwfiles_c2}{Path or array of paths to the bigWig files for second condition.}

\item{label_c1}{Condition name for first condition}

\item{label_c2}{Condition name for second condition}

\item{bin_size}{Bin size.}

\item{genome}{Genome. Available choices are mm9, hg38.}
}
\value{
a DESeqResults object as returned by DESeq2::results function
}
\description{
Runs a DESeq2 analysis on genome-wide bins of a specified bin size.
The particularity of this analysis is that it skips the estimateSizeFactors
step by default, because this is accounted for in the scaling step of
MINUTE-ChIP samples.
}
