% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{bw_bins_scatterplot}
\alias{bw_bins_scatterplot}
\title{Print a basic scatterplot from two bigwig files and an optional set of BED
files as highlighted annotations.}
\usage{
bw_bins_scatterplot(
  xtrack,
  ytrack,
  bsize = 10000,
  stat = "mean",
  genome = "mm9",
  highlight = NULL,
  minoverlap = 0L,
  highlight_label = NULL,
  xlimits = NULL,
  ylimits = NULL,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{xtrack}{BigWig file to be used for the x axis}

\item{ytrack}{BigWig file to be used for the y axis}

\item{bsize}{Bin size. Default 10000.}

\item{stat}{Statistics used in the calculation of bins.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{highlight}{List of bed files to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted}

\item{highlight_label}{Labels for the highlight groups.}

\item{xlimits}{Limits used for x axis}

\item{ylimits}{Limits used for y axis}

\item{xlabel}{Label for x axis}

\item{ylabel}{Label used for y axis
If not provided, filenames are used.}
}
\value{
A ggplot object
}
\description{
Print a basic scatterplot from two bigwig files and an optional set of BED
files as highlighted annotations.
}
