% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{bw_bins_violinplot}
\alias{bw_bins_violinplot}
\title{Violin plot of bin distribution of a set of bigWig files overlayed with
annotated bins (i.e. bins overlapping a given BED file)}
\usage{
bw_bins_violinplot(
  bw,
  bg_bw = NULL,
  bw_label = NULL,
  bsize = 10000,
  stat = "mean",
  genome = "mm9",
  highlight = NULL,
  minoverlap = 0L,
  norm_func = identity
)
}
\arguments{
\item{bw}{BigWig files}

\item{bg_bw}{BigWig files used as background (us. input)}

\item{bw_label}{Labels to use for in the plot for the bw files.}

\item{bsize}{Bin size. Default 10000.}

\item{stat}{Statistics used in the calculation of bins.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{highlight}{Bed file to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted}

\item{norm_func}{Function to use on top of dividing bw/bg_bw (usually identity or log2)}
}
\value{
A ggplot object
}
\description{
Violin plot of bin distribution of a set of bigWig files overlayed with
annotated bins (i.e. bins overlapping a given BED file)
}
