% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{bw_granges_diff_analysis}
\alias{bw_granges_diff_analysis}
\title{Compute DESeq2 differential analysis on GRanges objects}
\usage{
bw_granges_diff_analysis(
  granges_c1,
  granges_c2,
  label_c1,
  label_c2,
  estimate_size_factors = FALSE
)
}
\arguments{
\item{granges_c1}{GRanges object containing the values for condition 1.}

\item{granges_c2}{GRanges object containing the values for condition 2.
Note that these objects must correspond to the same loci.}

\item{label_c1}{Condition name for condition 1.}

\item{label_c2}{Condition name for condition 2.}

\item{estimate_size_factors}{If TRUE, normal DESeq2 procedure is done. Set it
to true to analyze non-MINUTE data.}
}
\value{
a DESeqResults object as returned by DESeq2::results function
}
\description{
Runs a DESeq2 analysis on loci specified on GRanges objects.
The particularity of this analysis is that it skips the estimateSizeFactors
step by default, because this is accounted for in the scaling step of
MINUTE-ChIP samples.
}
