% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{bw_profile_plot}
\alias{bw_profile_plot}
\title{Profile plot of a set of bw files
annotated bins (i.e. bins overlapping a given BED file)}
\usage{
bw_profile_plot(
  bw,
  bed,
  bg_bw = NULL,
  mode = "stretch",
  bsize = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = F,
  show_error = F,
  norm_func = identity
)
}
\arguments{
\item{bw}{BigWig files}

\item{bed}{BED file to use to summarize. It needs to have an adequate `name` field
(where names correspond to categories that can be grouped).}

\item{bg_bw}{BigWig files used as background (us. input)}

\item{mode}{How to align BED loci: Can be stretch, start, end, center.}

\item{bsize}{Bin size.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{ignore_strand}{Ignore strand information in BED file (default false).}

\item{show_error}{Show standard error.}

\item{norm_func}{Function to apply after bw / bg (default identity)}
}
\value{
A plot object
}
\description{
Profile plot of a set of bw files
annotated bins (i.e. bins overlapping a given BED file)
}
