% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{calculate_breakslist}
\alias{calculate_breakslist}
\title{Calculate breaks list for the values in a matrix}
\usage{
calculate_breakslist(values)
}
\arguments{
\item{values}{Value matrix.}
}
\value{
Sequence of breaks used by heatmap function
}
\description{
Given a numerical matrix, calculate breakslist to use in the summarized
heatmap plot. This is done in order to unify color scale regardless of
normalization used (mostly log vs linear), keeping white as the zero value.
}
\details{
This function ignores NA values to calculate min and max values.
}
