% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{calculate_bw_profile}
\alias{calculate_bw_profile}
\title{Calculate profile values of a bigWig file over GRanges object.}
\usage{
calculate_bw_profile(
  bw,
  granges,
  bg_bw = NULL,
  label = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = F,
  norm_func = identity
)
}
\arguments{
\item{bw}{BigWig file to be summarized.}

\item{granges}{GRanges object}

\item{bg_bw}{BigWig file to be used as background.}

\item{label}{Name to give to the values}

\item{mode}{How to handle differences in lengths across loci:

  stretch: Anchor each locus on both sides.

  start: Anchor all loci on start.

  end: Anchor all loci on end.

  center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}
}
\value{
A DataFrame with the aggregated scores
}
\description{
Calculates profile values of a bigWig file over the loci speficied in a
GRanges object. Data points are taken each bin_size base pairs.
}
\details{
Loci are aligned depending on mode parameter:

- stretch. Aligns all starts and all ends, sort of stretching the loci.
The median of these lenghts is taken as the pseudo-length in order to show
a realistic plot when displayed.

- start. All loci are aligned by start.

- end. All loci are aligned by end.

- center. All loci are aligned by center.

Adapted from seqplots rtracklayer wrapping functions to compute coverage values:
For more on seqplots: https://www.bioconductor.org/packages/release/bioc/html/seqplots.html
}
