% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{compute_breakslist}
\alias{compute_breakslist}
\title{Given a value matrix, compute breakslist for the summarized heatmap.}
\usage{
compute_breakslist(mat)
}
\arguments{
\item{mat}{Value matrix}
}
\value{
Sequence of breaks used by heatmap function
}
\description{
This is done so log scale and norm scale heatmaps have the same color scale
on the positive values, and zero values are shown as white always.
}
\details{
This function ignores NA values to calculate min and max values.
}
