% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{compute_stretch_matrix}
\alias{compute_stretch_matrix}
\title{Calculate matrix for stretch mode}
\usage{
compute_stretch_matrix(
  bw,
  gr,
  bin = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = F
)
}
\arguments{
\item{bw}{BigWigFile object}

\item{gr}{GRanges object}

\item{bin}{Bin size}

\item{upstream}{Number of basepairs upstream}

\item{downstream}{Number of basepairs downstream}

\item{ignore_strand}{Ignore strand (bool)}
}
\value{
Summary matrix
}
\description{
Calculate matrix for stretch mode
}
