% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{get_nreads_columns}
\alias{get_nreads_columns}
\title{Get values in a data frame object as round numeric values in a matrix.}
\usage{
get_nreads_columns(df, length_factor = 1000)
}
\arguments{
\item{df}{Target data frame}

\item{length_factor}{Scaling factor to multiply coverage values by.}
}
\value{
An integer matrix
}
\description{
This is an auxiliary function for stats. It drops NAs or NaN values, only
complete cases are used.
}
