% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot2-resembling default palette for an arbitrary number of values.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of values to be generated.}
}
\value{
A list of n colors.
}
\description{
ggplot2-resembling default palette for an arbitrary number of values.
}
