% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{granges_cbind}
\alias{granges_cbind}
\title{Performs a cbind operation on a GRanges list, appending scores}
\usage{
granges_cbind(grlist, labels)
}
\arguments{
\item{grlist}{A list of GRanges objects that have all the same fields.}

\item{labels}{Vector of names for the score columns.}
}
\description{
It will sort the GRanges elements in order to ensure the match is proper.
}
