% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{make_averages_df}
\alias{make_averages_df}
\title{Compute averages and standard error values for a matrix returned by
summary_matrix.}
\usage{
make_averages_df(matrix, label)
}
\arguments{
\item{matrix}{A matrix returned by summary_matrix}

\item{label}{Label for the sample}
}
\value{
A dataframe with summarized values, sderror and medians, plus a label.
}
\description{
Compute averages and standard error values for a matrix returned by
summary_matrix.
}
