% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{make_norm_label}
\alias{make_norm_label}
\title{Construct a string that represents the normalization function.}
\usage{
make_norm_label(f, bg)
}
\arguments{
\item{f}{Function name used to norm}

\item{bg}{Background}
}
\value{
A string with the corresponding label
}
\description{
Construct a string that represents the normalization function.
}
