% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normGeneCounts.R
\name{mean_ratio_norm}
\alias{mean_ratio_norm}
\title{Median Ratio Normalization of Gene Count Matrix}
\usage{
mean_ratio_norm(gene.counts)
}
\arguments{
\item{gene.counts}{Gene count table for normalization}
}
\value{
Normalised Counts
}
\description{
This function devides original values with median ratios comparing to the first sample
according to edgeR method.
}
