% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{multi_bw_ranges}
\alias{multi_bw_ranges}
\title{Intersect a list of bw files with a GRanges object}
\usage{
multi_bw_ranges(
  bwfiles,
  labels,
  granges,
  per_locus_stat = "mean",
  selection = NULL
)
}
\arguments{
\item{bwfiles}{BigWig file list to be summarized.}

\item{labels}{Names to be assigned to the columns}

\item{granges}{GRanges object to intersect}

\item{per_locus_stat}{Aggregating function per stat}

\item{selection}{A GRanges object to restrict analysis to.}
}
\value{
a sorted GRanges object
}
\description{
Build a binned-scored GRanges object from a list of bigWig files. The
aggregating function per locus can be min, max, sd, mean.
}
