% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{natural_sort_by_field}
\alias{natural_sort_by_field}
\title{Moves a column of a dataframe to rownames and naturally sorts the rows}
\usage{
natural_sort_by_field(df, col)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The column (usually name)}
}
\value{
A sorted df
}
\description{
Moves a column of a dataframe to rownames and naturally sorts the rows
}
