% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{norm_col}
\alias{norm_col}
\title{Normalize a column of a dataframe, replacing it.}
\usage{
norm_col(df, fg_col, bg_col, norm_func)
}
\arguments{
\item{df}{Data frame to be used.}

\item{fg_col}{Column that corresponds to main signal.}

\item{bg_col}{Column that corresponds to background.}

\item{norm_func}{Function to apply after dividing.}
}
\description{
Normalize a column of a dataframe, replacing it.
}
