% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{palette_categorical}
\alias{palette_categorical}
\title{Palette for qualitative datasets.}
\usage{
palette_categorical(n)
}
\arguments{
\item{n}{Number of values (must be <=12.}
}
\value{
A list of n colors.
}
\description{
Returns a palette useful for qualitative datasets.
Maximum value of 12 colors.
}
