% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bed_summary_heatmap}
\alias{plot_bw_bed_summary_heatmap}
\title{Summary heatmap of a categorized BED file}
\usage{
plot_bw_bed_summary_heatmap(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  labels = NULL,
  aggregate_by = "true_mean",
  norm_func = identity,
  file_out = NA
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bedfile}{BED file to intersect with the BigWig file.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{Labels to use for in the plot for the bw files.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}

\item{file_out}{Output the plot to a file.}
}
\value{
A pheatmap object
}
\description{
Make a summary heatmap where each cell contains an aggregated value of a
bigWig file from bwfiles and a category of a BED file (bedfile). The
provided BED file must have a name field that is valid (i.e. can be grouped,
representing some type of category, not a per-locus unique ID).
}
