% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bins_diff_scatter}
\alias{plot_bw_bins_diff_scatter}
\title{Scatter plot with significance testing.}
\usage{
plot_bw_bins_diff_scatter(
  bwfiles_c1,
  bwfiles_c2,
  label_c1,
  label_c2,
  bg_bwfiles_c1 = NULL,
  bg_bwfiles_c2 = NULL,
  bin_size = 10000,
  genome = "mm9",
  estimate_size_factors = FALSE,
  pval_cutoff = NULL,
  logfc_cutoff = NULL,
  norm_func_x = identity,
  norm_func_y = identity
)
}
\arguments{
\item{bwfiles_c1}{Condition 1 bigwig files.}

\item{bwfiles_c2}{Condition 2 bigwig files.}

\item{label_c1}{Label for condition 1.}

\item{label_c2}{Label for condition 2.}

\item{bg_bwfiles_c1}{Optional condition 1 bigwig input (background) files.}

\item{bg_bwfiles_c2}{Optional condition 2 bigwig input (background) files.}

\item{bin_size}{Bin size.}

\item{genome}{Genome used. Must be one of supported.}

\item{estimate_size_factors}{If True, the estimateSizeFactors step from
DESeq2 is not skipped.}

\item{pval_cutoff}{P-value cutoff.}

\item{logfc_cutoff}{Log fold change cutoff. If logfc_cutoff is negative,
bins under logfc_cutoff will be highlighted.}

\item{norm_func_x}{Function to use after x / x_bg.}

\item{norm_func_y}{Function to use after y / y_bg.}
}
\value{
ggplot object
}
\description{
Make a scatterplot where bins that pass a specific statistical significance
threshold across replicates are highlighted.
}
