% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bins_diff_volcano}
\alias{plot_bw_bins_diff_volcano}
\title{Volcano plot for differential analysis of genome-wide bins}
\usage{
plot_bw_bins_diff_volcano(
  bwfiles_c1,
  bwfiles_c2,
  label_c1,
  label_c2,
  bin_size = 10000,
  genome = "mm9",
  estimate_size_factors = FALSE,
  pval_cutoff = NULL,
  logfc_cutoff = NULL
)
}
\arguments{
\item{bwfiles_c1}{Condition 1 bigwig files.}

\item{bwfiles_c2}{Condition 2 bigwig files.}

\item{label_c1}{Label for condition 1.}

\item{label_c2}{Label for condition 2.}

\item{bin_size}{Bin size.}

\item{genome}{Genome used. Must be one of supported.}

\item{estimate_size_factors}{If True, the estimateSizeFactors step from
DESeq2 is not skipped.}

\item{pval_cutoff}{P-value cutoff.}

\item{logfc_cutoff}{Log fold change cutoff. In volcano plot this cutoff is
used as absolute value (-logfc cutoff and + logfc cutoff are selected).}
}
\value{
ggplot object
}
\description{
Make a volcano plot where bins that pass a specific statistical significance
threshold across replicates are highlighted.
}
