% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normGeneCounts.R
\name{quantile_norm}
\alias{quantile_norm}
\title{Quantile Normalization of Gene Count Matrix}
\usage{
quantile_norm(gene.counts)
}
\arguments{
\item{gene.counts}{Gene count table for normalization}
}
\value{
Normalised Counts
}
\description{
This function replaces original values with estimated values according
to the rank within each sample.
}
