% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{rename_score}
\alias{rename_score}
\title{Rename score function of a GRanges object}
\usage{
rename_score(gr, new.name)
}
\arguments{
\item{gr}{GRanges object}

\item{new.name}{Name to give to the column.}
}
\description{
Rename score function of a GRanges object
}
