% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{render_bw_report}
\alias{render_bw_report}
\title{Renders a report with specified bigwig directory and bedfiles}
\usage{
render_bw_report(
  bwdir,
  bed,
  outfile = "bwreport.html",
  highlight = NULL,
  bsize = 10000,
  minoverlap = 2000,
  genome = "mm9",
  outdir = ".",
  force = FALSE
)
}
\arguments{
\item{bwdir}{Directory where bigwig files are}

\item{bed}{Bed file that contains some regions of interest to summarize}

\item{outfile}{Output filename (without the full path)}

\item{highlight}{Optional bed file to highlight bins overlapping (bin analysis)}

\item{bsize}{Bin size used.}

\item{minoverlap}{Minimum overlap between a bin and a highlighted region to be marked.}

\item{genome}{Genome used (supported: mm9, hg38)}

\item{outdir}{Output directory}

\item{force}{Force creation of tables or use them if they exist.}
}
\description{
Renders a report with specified bigwig directory and bedfiles
}
