% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normGeneCounts.R
\name{size_factor_norm}
\alias{size_factor_norm}
\title{Size Factor Normalization of Gene Count Matrix}
\usage{
size_factor_norm(gene.counts)
}
\arguments{
\item{gene.counts}{Gene count table for normalization}
}
\value{
Normalised Counts
}
\description{
This function devides original values with size factors of each sample according
to DESeq method.
}
