% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{summary_heatmap}
\alias{summary_heatmap}
\title{Plot a pretty heatmap using pheatmap library}
\usage{
summary_heatmap(values, title, size = 35, file_out = NA)
}
\arguments{
\item{values}{Value matrix}

\item{title}{Title of the plot}

\item{size}{Size of the square}

\item{file_out}{Optionally save directly to a file. This will take care of
cropping the heatmap to a right size so it fits.}
}
\value{
Heatmap plot
}
\description{
This function ignores NA values to calculate min and max values.
}
