% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{summary_matrix}
\alias{summary_matrix}
\title{Summarize a BigWig file over loci on a GRanges object.}
\usage{
summary_matrix(track, gr, npoints, ignore_strand = F)
}
\arguments{
\item{track}{BigWigFile object (rtracklayer)}

\item{gr}{GRanges object}

\item{npoints}{How many points to take}

\item{ignore_strand}{Ignore strand information in gr. Default false.}
}
\value{
A value matrix dimensions len(gr) x npoints
}
\description{
Summarize a BigWig file over loci on a GRanges object.
}
