% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_elsasserlab_print}
\alias{theme_elsasserlab_print}
\title{Theme for print versions of plots.}
\usage{
theme_elsasserlab_print(
  base_size = 22,
  base_family = "",
  base_line_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size. Size of elements in the plot are calculated
from this.}

\item{base_family}{Font family.}

\item{base_line_size}{Thickness of lines. Default is defined based on base_size.}
}
\description{
A theme where font sizes will generally be pretty large compared to screen
values, but good to show in combined panels.
}
