% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_elsasserlab_screen}
\alias{theme_elsasserlab_screen}
\title{Theme for screen plotting.}
\usage{
theme_elsasserlab_screen(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size. Size of elements in the plot are calculated
from this.}

\item{base_family}{Font family.}

\item{base_line_size}{Thickness of lines. Default is defined based on base_size.}
}
\description{
A theme for using on screen or single plots that plays nice with the print
version.
}
