% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normGeneCounts.R
\name{trim_quantile}
\alias{trim_quantile}
\title{Limit Outliers of Gene Count Matrix}
\usage{
trim_quantile(x, q = 0.995)
}
\arguments{
\item{x}{Gene count table or a vector}

\item{q}{Limit of quantile}
}
\value{
Trimmed Counts Matrix
}
\description{
This function trims extreme values above or below the indicated quantile, 
processes by column.
}
