% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{validate_categories}
\alias{validate_categories}
\title{Checks that the number of categories is reasonable for an aggregation.}
\usage{
validate_categories(cat_values)
}
\arguments{
\item{cat_values}{An array of values}
}
\description{
Throws a warning if found more than  50 values.
}
