% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{aggregate_scores}
\alias{aggregate_scores}
\title{Aggregate scores of a GRanges object on a specific field}
\usage{
aggregate_scores(scored.gr, group.col, aggregate.by)
}
\arguments{
\item{scored.gr}{A GRanges object with numerical metadata columns}

\item{group.col}{A column among the mcols that can be seen as a factor.}

\item{aggregate.by}{Function used to aggregate (mean, median, any valid
function).}
}
\value{
A DataFrame with the aggregated scores (any numerical column will be
    aggregated).
}
\description{
Aggregates scores of a GRanges object on a specific field.
}
