% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bed}
\alias{bw_bed}
\title{Build a scored GRanges object from a BED file.}
\usage{
bw_bed(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  labels = NULL,
  per_locus_stat = "mean",
  aggregate_by = NULL,
  norm_func = identity
)
}
\arguments{
\item{bwfiles}{BigWig file (or list) to be summarized.}

\item{bedfile}{BED file to intersect with the BigWig file.}

\item{bg_bwfiles}{BigWig file (or list) to be used as background to normalize to.
When aggregate_by is not NULL, the normalization to background happens
AFTER aggregating the values. For the opposite result, it is possible
to run bw_bed with aggregate_by = NULL and summarize manually after.}

\item{labels}{Column names of the score fields. Must have the same
length as bigwig file list. If not provided, labels are the names of
the files in bwfiles.}

\item{per_locus_stat}{Aggregate per locus function.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{norm_func}{Function to be applied after normalizing: norm_func(bw/bg).}
}
\description{
Build a scored GRanges object from a bigWig file and a BED file.
The aggregating function (per locus) can be min, max, sd, mean.
}
