% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bed}
\alias{bw_bed}
\title{Build a scored GRanges object from a BED file.}
\usage{
bw_bed(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  colnames = NULL,
  per.locus.stat = "mean",
  aggregate.by = NULL,
  norm_func = identity
)
}
\arguments{
\item{bwfiles}{BigWig file (or list) to be summarized.}

\item{bedfile}{BED file to intersect with the BigWig file.}

\item{colnames}{Column names of the score fields. Must have the same
length as bigwig file list. If not provided, colnames are the names of
the files in bwfiles.}

\item{per.locus.stat}{Aggregate per locus function.}

\item{aggregate.by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}
}
\description{
Build a scored GRanges object from a bigWig file and a BED file.
The aggregating function (per locus) can be min, max, sd, mean.
}
