% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{bw_bed_diff_analysis}
\alias{bw_bed_diff_analysis}
\title{Run DESeq2 analysis on bed file}
\usage{
bw_bed_diff_analysis(
  bwfiles_c1,
  bwfiles_c2,
  bedfile,
  label_c1,
  label_c2,
  estimate_size_factors = FALSE,
  keep_values = FALSE
)
}
\arguments{
\item{bwfiles_c1}{Path or array of paths to the bigWig files for first condition.}

\item{bwfiles_c2}{Path or array of paths to the bigWig files for second condition.}

\item{bedfile}{BED file for locus specific analysis.}

\item{label_c1}{Condition name for condition 1.}

\item{label_c2}{Condition name for condition 2.}

\item{estimate_size_factors}{If TRUE, normal DESeq2 procedure is done. Set it
to true to analyze non-MINUTE data.}

\item{keep_values}{If true, results also include the separate bin values.
This is used to avoid recalculation in plotting functions, but set to
false by default.}
}
\value{
a DESeqResults object as returned by DESeq2::results function
}
\description{
Runs a DESeq2 analysis on a set of loci specified in a BED file.
The particularity of this analysis is that it skips the estimateSizeFactors
step by default, because this is accounted for in the scaling step of
MINUTE-ChIP samples.
}
