% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{bw_bins_scatterplot}
\alias{bw_bins_scatterplot}
\title{Basic scatterplot from two bigwig files and an optional set of BED
files as highlighted annotations.}
\usage{
bw_bins_scatterplot(
  x,
  y,
  bg_x = NULL,
  bg_y = NULL,
  bsize = 10000,
  stat = "mean",
  genome = "mm9",
  highlight = NULL,
  minoverlap = 0L,
  highlight_label = NULL,
  norm_func_x = identity,
  norm_func_y = identity
)
}
\arguments{
\item{x}{BigWig file to be used for the x axis}

\item{y}{BigWig file to be used for the y axis}

\item{bg_x}{BigWig file to be used for the x axis background (us. input)}

\item{bg_y}{BigWig file to be used for the y axis background (us. input)}

\item{bsize}{Bin size. Default 10000.}

\item{stat}{Statistics used in the calculation of bins.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{highlight}{List of bed files to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted}

\item{highlight_label}{Labels for the highlight groups.
If not provided, filenames are used.}

\item{norm_func_x}{Function to use after x / x_bg}

\item{norm_func_y}{Function to use after y / y_bg}
}
\value{
A ggplot object
}
\description{
Basic scatterplot from two bigwig files and an optional set of BED
files as highlighted annotations.
}
