% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_profile}
\alias{bw_profile}
\title{Compute a coverage profile on a list of bw files with a given GRanges object.}
\usage{
bw_profile(
  bwfiles,
  bedfile,
  colnames,
  mode = "stretch",
  bin = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = F
)
}
\arguments{
\item{bwfiles}{BigWig file list to be summarized.}

\item{bedfile}{BED file to summarize}

\item{colnames}{Names to be assigned to the columns}
}
\value{
a data frame in long format
}
\description{
Compute a coverage profile on a list of bw files with a given GRanges object.
}
