% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{calculate_bw_profile}
\alias{calculate_bw_profile}
\title{Mean coverage of a BigWig list of files on a set of loci specified as BED file.}
\usage{
calculate_bw_profile(
  bw,
  gr,
  label = NULL,
  mode = "stretch",
  bin = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = F
)
}
\arguments{
\item{bw}{BigWig file to be summarized.}

\item{gr}{GRanges object}

\item{label}{Name to give to the values}

\item{mode}{How to handle differences in lengths across loci:
stretch: Anchor each locus on both sides.
start: Anchor all loci on start.
end: Anchor all loci on end.
center: Center all loci.}

\item{bin}{Bin size. Length of bin in base pairs. The lower, the higher the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{ignore_strand}{Whether to use strand information in BED file.}
}
\value{
A DataFrame with the aggregated scores
}
\description{
Adapted from seqplots rtracklayer wrapping functions to compute coverage values:
For more on seqplots: https://www.bioconductor.org/packages/release/bioc/html/seqplots.html
}
