% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwstats.R
\name{get_nreads_columns}
\alias{get_nreads_columns}
\title{Get values in a granges object as round numeric values in a matrix.}
\usage{
get_nreads_columns(granges, fraglen = 150)
}
\arguments{
\item{granges}{Target granges object}

\item{fraglen}{Estimated fragment length}
}
\value{
An integer matrix
}
\description{
This is an auxiliary function for stats. It drops NAs or NaN values, only
complete cases are used.
}
