% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{multi_bw_ranges}
\alias{multi_bw_ranges}
\title{Intersect a list of bw files with a GRanges object}
\usage{
multi_bw_ranges(
  bwfilelist,
  colnames,
  gr,
  per.locus.stat = "mean",
  selection = NULL
)
}
\arguments{
\item{bwfilelist}{BigWig file to be summarized.}

\item{colnames}{Names to be assigned to the columns}

\item{gr}{GRanges object to intersect}

\item{per.locus.stat}{Aggregating function per stat}

\item{selection}{A GRanges object to restrict analysis to.}
}
\description{
Build a binned-scored GRanges object from a list of bigWig files. The
aggregating function per locus can be min, max, sd, mean.
}
