% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{multi_bw_ranges}
\alias{multi_bw_ranges}
\title{Build a binned-scored GRanges object from a list of bigWig files}
\usage{
multi_bw_ranges(bwfilelist, col.names, gr, stat = "mean")
}
\arguments{
\item{stat}{Aggregating function. Mean by default. Choices: min, max, sd, mean.}

\item{bwfile}{BigWig file to be summarized.}

\item{bsize}{Bin size. Default 10000.}

\item{genome}{Genome. Available choices are mm9, hg38.}
}
\description{
Build a binned-scored GRanges object from a list of bigWig files. The aggregating
function can be min, max, sd, mean.
}
