% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{multi_bw_ranges_norm}
\alias{multi_bw_ranges_norm}
\title{Intersect a list of bw files with a GRanges object and normalize to a set
of background bwfiles.
#'}
\usage{
multi_bw_ranges_norm(
  bwfilelist,
  bg_bwfilelist,
  labels,
  granges,
  per_locus_stat = "mean",
  selection = NULL,
  norm_func = identity
)
}
\arguments{
\item{bwfilelist}{BigWig file list to be summarized.}

\item{bg_bwfilelist}{Background BigWig files.}

\item{labels}{Names to be assigned to the columns}

\item{granges}{GRanges object to intersect}

\item{per_locus_stat}{Aggregating function per stat}

\item{selection}{A GRanges object to restrict analysis to.}

\item{norm_func}{Function to apply after bw/bg_bw.}
}
\value{
a sorted GRanges object
}
\description{
Intersect a list of bw files with a GRanges object and normalize to a set
of background bwfiles.
#'
}
