% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bed_summary_heatmap}
\alias{plot_bw_bed_summary_heatmap}
\title{Violin plot of bin distribution of a set of bigWig files overlayed with
annotated bins (i.e. bins overlapping a given BED file)}
\usage{
plot_bw_bed_summary_heatmap(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  labels = NULL,
  aggregate_by = "true_mean",
  norm_func = identity,
  file_out = NA
)
}
\arguments{
\item{bwfiles}{BigWig files.}

\item{bedfile}{BED file to use to summarize. It needs to have an adequate `name` field
(where names correspond to categories that can be grouped).}

\item{bg_bwfiles}{BigWig files used as background (us. input).}

\item{labels}{Labels to use for in the plot for the bw files.}

\item{aggregate_by}{Can be true_mean, mean (mean of means), median (median of means).}

\item{norm_func}{Function to use on top of dividing bw/bg_bw (usually identity or log2).}

\item{file_out}{Output the plot to a file.}
}
\value{
A plot object
}
\description{
Violin plot of bin distribution of a set of bigWig files overlayed with
annotated bins (i.e. bins overlapping a given BED file)
}
