% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_profile}
\alias{plot_bw_profile}
\title{Profile plot of a set of bigWig files}
\usage{
plot_bw_profile(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  ignore_strand = FALSE,
  show_error = FALSE,
  norm_func = identity
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bedfile}{BED file to summarize}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{mode}{How to handle differences in lengths across loci:

  stretch: Anchor each locus on both sides.

  start: Anchor all loci on start.

  end: Anchor all loci on end.

  center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{show_error}{Show standard error.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}
}
\value{
A ggplot object.
}
\description{
Plots a profile of a set of bigWig files over a set of loci in a BED file.
}
