% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{summary_heatmap}
\alias{summary_heatmap}
\title{Plot a pretty heatmap using pheatmap library}
\usage{
summary_heatmap(values, title, size = 35, file_out = NA)
}
\arguments{
\item{values}{Value matrix.}

\item{title}{Plot title.}

\item{size}{Square size in points.}

\item{file_out}{Optional file output.}
}
\value{
pheatmap object
}
\description{
This function ignores NA values to calculate min and max values.
}
